Extract the VB.Net hack into a separate helper
